Metadata refactoring revision: removing clause that manually copied layout data from a composite media element onto its inner layout target sprite. Instead, passing the media element's layout metadata as the layout target sprite's constructor, so that both use the same instance, making sure that mutations on the instance reach the layout target sprite.
NOTE that as a result of the new metadata approach, the user can be overriding the set layoutMetadata instance on media element. When this happens, the inner layout target sprite will no longer be pointing to the correct layout metadata instance. Fixing this requires changing layoutTargetSprite to be capable of dealing with a variable layoutMetadata instance (for right now the class assumes the instance to be static, either passed in at construction time, or constructed by the class itself at that time).

git-svn-id: http://opensource.adobe.com/svn/opensource/osmf/trunk@14596 a9308255-753e-0410-a2e9-80b3fbc4fff6