Fix bug FM-853.
The fix is different from what is described in the report. Instead of changing the behavior of MediaPlayer, I decided to change when HTTPNetStream dispatches the play event such that it conforms to the behavior of regular NetStream.

During the course, I ran into an issue with OSMF integration test regarding the tests for http streaming. I have tried a few things but the problem persistent. Please refer to the comments in OSMFIntegrationTests for detail. In a gist, I decide to comment out the tests that causes the failure, check in the fix and revisit the issue at a later time.

git-svn-id: http://opensource.adobe.com/svn/opensource/osmf/trunk@16657 a9308255-753e-0410-a2e9-80b3fbc4fff6